Move expansion to serializers, not search.
