+++ aleph/search/result.py
@@ -6,18 +6,9 @@ from aleph.index.util import unpack_result
 from aleph.search.parser import QueryParser
 from aleph.search.facet import CategoryFacet, CollectionFacet, CountryFacet
 from aleph.search.facet import LanguageFacet, SchemaFacet, Facet
-from aleph.search.expand import RoleExpander, DocumentExpander
-from aleph.search.expand import EntitiesExpander, CollectionExpander
 
 
 class QueryResult(object):
-    EXPANDERS = [
-        CollectionExpander('collection_id', 'collection'),
-        EntitiesExpander('entities', 'related'),
-        DocumentExpander('parent', 'parent'),
-        RoleExpander('uploader_id', 'uploader'),
-        RoleExpander('creator', 'creator'),
-    ]
 
     def __init__(self, request, parser=None, results=None, total=None,
                  schema=None):
@@ -41,23 +32,13 @@ class QueryResult(object):
         args.extend(self.parser.items)
         return self.request.base_url + query_string(args)
 
-    def process(self):
-        results = []
-        for result in self.results:
-            for expander in self.EXPANDERS:
-                expander.collect(self, result)
-            results.append(result)
-
-        for result in results:
-            for expander in self.EXPANDERS:
-                expander.apply(self, result)
-            data, err = self.schema().dump(result)
-            yield data
-
     def to_dict(self):
+        results = list(self.results)
+        if self.schema:
+            results, _ = self.schema().dump(results, many=True)
         return {
             'status': 'ok',
-            'results': list(self.process()),
+            'results': results,
             'total': self.total,
             'page': self.parser.page,
             'limit': self.parser.limit,
